+++ helpers/inbound/inbound.go
@@ -3,7 +3,6 @@ package inbound
 import (
 	"io"
 	"io/ioutil"
-	"log"
 	"mime"
 	"mime/multipart"
 	"net/http"
@@ -17,21 +16,21 @@ type ParsedEmail struct {
 	rawRequest  *http.Request
 }
 
-func Parse(request *http.Request) *ParsedEmail {
+func Parse(request *http.Request) (*ParsedEmail, error) {
 	result := ParsedEmail{
 		Headers:     make(map[string]string),
 		Body:        make(map[string]string),
 		Attachments: make(map[string][]byte),
 		rawRequest:  request,
 	}
-	result.parse()
-	return &result
+	err := result.parse()
+	return &result, err
 }
 
-func (email *ParsedEmail) parse() {
+func (email *ParsedEmail) parse() error {
 	err := email.rawRequest.ParseMultipartForm(0)
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 	emails := email.rawRequest.MultipartForm.Value["email"]
 	headers := email.rawRequest.MultipartForm.Value["headers"]
@@ -41,18 +40,27 @@ func (email *ParsedEmail) parse() {
 	if len(emails) > 0 {
 		email.parseRawEmail(emails[0])
 	}
+
+	return nil
 }
 
-func (email *ParsedEmail) parseRawEmail(rawEmail string) {
+func (email *ParsedEmail) parseRawEmail(rawEmail string) error {
 	sections := strings.SplitN(rawEmail, "\n\n", 2)
 	email.parseHeaders(sections[0])
-	raw := parseMultipart(strings.NewReader(sections[1]), email.Headers["Content-Type"])
+	raw, err := parseMultipart(strings.NewReader(sections[1]), email.Headers["Content-Type"])
+	if err != nil {
+		return err
+	}
+
 	for {
 		emailPart, err := raw.NextPart()
 		if err == io.EOF {
-			return
+			return nil
 		}
-		rawEmailBody := parseMultipart(emailPart, emailPart.Header.Get("Content-Type"))
+		rawEmailBody, err := parseMultipart(emailPart, emailPart.Header.Get("Content-Type"))
+		if err != nil {
+			return err
+		}
 		if rawEmailBody != nil {
 			for {
 				emailBodyPart, err := rawEmailBody.NextPart()
@@ -60,36 +68,40 @@ func (email *ParsedEmail) parseRawEmail(rawEmail string) {
 					break
 				}
 				header := emailBodyPart.Header.Get("Content-Type")
-				email.Body[header] = string(readBody(emailBodyPart))
+				b, err := ioutil.ReadAll(emailPart)
+				if err != nil {
+					return err
+				}
+				email.Body[header] = string(b)
 			}
 
 		} else if emailPart.FileName() != "" {
-			email.Attachments[emailPart.FileName()] = readBody(emailPart)
+			b, err := ioutil.ReadAll(emailPart)
+			if err != nil {
+				return err
+			}
+			email.Attachments[emailPart.FileName()] = b
 		} else {
 			header := emailPart.Header.Get("Content-Type")
-			email.Body[header] = string(readBody(emailPart))
+			b, err := ioutil.ReadAll(emailPart)
+			if err != nil {
+				return err
+			}
+			email.Body[header] = string(b)
 		}
 	}
 }
 
-func parseMultipart(body io.Reader, contentType string) *multipart.Reader {
+func parseMultipart(body io.Reader, contentType string) (*multipart.Reader, error) {
 	mediaType, params, err := mime.ParseMediaType(contentType)
 	if err != nil {
-		log.Fatal(err)
+		return nil, err
 	}
 
 	if strings.HasPrefix(mediaType, "multipart/") {
-		return multipart.NewReader(body, params["boundary"])
+		return multipart.NewReader(body, params["boundary"]), nil
 	}
-	return nil
-}
-
-func readBody(body io.Reader) []byte {
-	raw, err := ioutil.ReadAll(body)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return raw
+	return nil, nil
 }
 
 func (email *ParsedEmail) parseHeaders(headers string) {