+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+[2021-02-10] Version 3.8.0
+--------------------------
+**Library - Fix**
+- [PR #426](https://github.com/sendgrid/sendgrid-go/pull/426): typo in method name. Thanks to [@thinkingserious](https://github.com/thinkingserious)!
+- [PR #355](https://github.com/sendgrid/sendgrid-go/pull/355): content value issue by implementing NewSingleEmailPlanText. Thanks to [@prakashpandey](https://github.com/prakashpandey)!
+- [PR #398](https://github.com/sendgrid/sendgrid-go/pull/398): Add error handling for upstream on inbound parse. Thanks to [@thavanle](https://github.com/thavanle)!
+
+**Library - Feature**
+- [PR #425](https://github.com/sendgrid/sendgrid-go/pull/425): Add support for more bypass settings. Thanks to [@yousifh](https://github.com/yousifh)!
+
+
 [2020-11-18] Version 3.7.2
 --------------------------
 **Library - Docs**